+++ src/Compute/Compute/ChangeLog.md
@@ -28,6 +28,7 @@
 * Fixed the `New-AzVM` bug to avoid accessing the `EncryptionAtHost` property for subscriptions who cannot access it since it is behind a feature flag.
 * Updated `Get-AzVmExtension` to return instanceView when used with `-Status`.
 * Reverted SSH Private Key File permission changes in `New-AzVm`.
+* Fixed `New-AzVmss` to correctly work when using `-EdgeZone` by creating the Load Balancer in the correct edge zone. 
 
 ## Version 7.0.0
 * Added update functionality in `Update-AzVmss` for parameters `SecurityType`, `EnableSecureBoot`, and `EnableVtpm` for the parameter set with the Put operation.